Split primitives.js into separate modules
